# backend/docs/swagger.yaml
basePath: /
definitions:
  models.Exam:
    properties:
      created_at:
        type: string
      department_id:
        type: string
      file:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      semester:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      uploaded_by:
        type: string
      year:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: IKnow Learning Platform API
  title: IKnow API
  version: "1.0"
paths:
...
  /api/exams:
    get:
      consumes:
      - application/json
      description: Retrieve exams for the logged-in user with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get paginated exams
      tags:
      - Exams
    post:
      consumes:
      - multipart/form-data
      description: Upload a PDF exam with title, year, semester, and department ID
      parameters:
      - description: Title of the exam
        in: formData
        name: title
        required: true
        type: string
      - description: Year of study
        in: formData
        name: year
        required: true
        type: integer
      - description: Semester (1 or 2)
        in: formData
        name: semester
        required: true
        type: integer
      - description: Department ID
        in: formData
        name: department_id
        required: true
        type: string
      - description: PDF file to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Exam'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new exam
      tags:
      - Exams
  /api/exams/{id}:
    get:
      description: Returns a exam by its ID
      parameters:
      - description: Exam ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Exam'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a exam by ID
      tags:
      - Exams
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"